refactor: rename, backtop -> back-top
