--- packages/back-top/CHANGELOG.md
@@ -1,1 +1,0 @@
-# @illa-design/BackTop
