--- packages/back-top/README.md
@@ -1,25 +1,0 @@
-# BackTop
-
-`BackTop` makes it easy to go back to the top of the page.
-
-## Installation
-
-```bash
-yarn add @illa-design/backtop
-```
-
-## Import component
-
-```jsx
-import { Button } from "@illa-dedign/backtop"
-```
-
-## API
-
-### BackTop Basic Properties
-
-| Props     | Desc                                                  | Type                                                         | Default   |
-| ----------- | ----------------------------------------------------- | ------------------------------------------------------------ | -------- |
-
-
-## Example
