--- packages/back-top/README.zh-CN.md
@@ -1,27 +1,0 @@
-# 返回顶部 BackTop
-
-可一键返回顶部的按钮
-
-## 安装
-
-```bash
-yarn add @illa-design/backtop
-```
-
-## 引用组件
-
-```jsx
-import { BackTop } from "@illa-dedign/backtop"
-```
-
-## 组件接口(API)
-
-### BackTop 基础属性
-
-| 参数名      | 描述                         | 类型                                                         | 默认值   |
-| ----------- | ---------------------------- | ------------------------------------------------------------ | -------- |
-
-
-## 使用方法
-
-### 基础用法
