+++ packages/back-top/src/index.ts
@@ -1,3 +1,3 @@
-export * from "./backtop"
+export * from "./back-top"
 export * from "./interface"
 export * from "./style"
