+++ packages/back-top/vite.config.ts
@@ -29,7 +29,7 @@ export default defineConfig({
     minify: true,
     lib: {
       entry: path.resolve(__dirname, "src/index.ts"),
-      name: "@illa-design/backtop",
+      name: "@illa-design/back-top",
       fileName: (format) => `${format}/index.js`,
     },
     rollupOptions: {
